+++ src/app/(store)/layout.tsx
@@ -1,4 +1,4 @@
-import { ReactNode, Suspense } from 'react'
+import { ReactNode } from 'react'
 
 import { Header } from '@/components/header'
 import { CartProvider } from '@/contexts/cart-context'
@@ -7,10 +7,8 @@ export default function StoreLayout({ children }: { children: ReactNode }) {
   return (
     <CartProvider>
       <div className="mx-auto grid min-h-screen w-full max-w-[1440px] grid-rows-app gap-5 p-8">
-        <Suspense>
-          <Header />
-          {children}
-        </Suspense>
+        <Header />
+        {children}
       </div>
     </CartProvider>
   )
